+++ R/RunPlpHelpers.R
@@ -34,10 +34,17 @@ checkInputs <- function(inputs) {
     )
     
     # check class is correct
-    if(class(inputs[[inputName]]) != inputName){
+    if(class(inputs[[inputName]]) != inputName && class(inputs[[inputName]]) != 'list'){
       ParallelLogger::logError(paste0('Incorrect ', inputName))
       stop('Bad input')
-    }
+    } 
+    
+    if(class(inputs[[inputName]]) == 'list'){
+      if(unique(unlist(lapply(inputs[[inputName]], class))) != inputName){
+        ParallelLogger::logError(paste0('Incorrect ', inputName))
+        stop('Bad input list')
+      }
+    } 
     
   }