+++ inst/shiny/PLPViewer/modules/settings.R
@@ -80,31 +80,26 @@ formatModSettings <- function(modelSettings){
 
 # format covariateSettings
 formatCovSettings <- function(covariateSettings){
-  if(class(covariateSettings)=='list'){
-    #code for when multiple covariateSettings
-    covariates <- c() 
-    for(i in 1:length(covariateSettings)){
-      if(attr(covariateSettings[[i]],'fun')=='getDbDefaultCovariateData'){
-        covariatesTemp <- data.frame(covariateName = names(covariateSettings[[i]]), 
-                                     SettingValue = unlist(lapply(covariateSettings[[i]], 
-                                                                  function(x) paste0(x, 
-                                                                                     collapse='-'))))
-      } else{
-        covariatesTemp <- data.frame(covariateName = covariateSettings[[i]]$covariateName,
-                                     SettingValue = ifelse(sum(names(covariateSettings[[i]])%in%c("startDay","endDay"))>0,
-                                                           paste(names(covariateSettings[[i]])[names(covariateSettings[[i]])%in%c("startDay","endDay")],
-                                                                 covariateSettings[[i]][names(covariateSettings[[i]])%in%c("startDay","endDay")], sep=':', collapse = '-'),
-                                                           "")
+  
+  if(class(covariateSettings)=='covariateSettings'){
+    covariateSettings <- list(covariateSettings)
+  }
+  
+  #code for when multiple covariateSettings
+  covariates <- c() 
+  for(i in 1:length(covariateSettings)){
+    covariatesTemp <- data.frame(
+      fun = attr(covariateSettings[[i]],'fun'),
+      setting = i,
+      covariateName = names(covariateSettings[[i]]), 
+      SettingValue = unlist(
+        lapply(
+          covariateSettings[[i]], 
+          function(x) paste0(x, collapse='-')
         )
-        
-      }
-      covariates  <- rbind(covariates,covariatesTemp)
-    }
-  } else{
-    covariates <- data.frame(covariateName = names(covariateSettings), 
-                             SettingValue = unlist(lapply(covariateSettings, 
-                                                          function(x) paste0(x, 
-                                                                             collapse='-'))))
+      )
+    )
+    covariates  <- rbind(covariates,covariatesTemp)
   }
   row.names(covariates) <- NULL
   return(covariates)