+++ tests/testthat/test-helperfunctions_updated.R
@@ -46,5 +46,12 @@ test_that("setPythonEnvironment", {
   testthat::expect_equal(class(setPythonEnvironment(envname='madeup34343', envtype = 'conda')), "character")
 })
 
+test_that("Borrowed cut2", {
+   x <- c(1, rep(2, 2), rep(4, 4), rep(5, 5), rep(6, 6)) 
+   groups <- PatientLevelPrediction:::cut2(x, g = 3)
+   expect_true(
+     all(levels(groups) == c("[1,5)", "5", "6"))
+   )
+})
 
 # getOs test?